requirements.txt
Flask 
SQLAlchemy 
pymysql 
Flask-Migrate
python-dotenv
Flask-JWT-Extended
Werkzeug
